+++ frontend/components/lines/LinesListViewAll/index.tsx
